Criado métodos para salvar um livro como alugado por um usuário, validando se ele já possui aquele livro alugado, se ele já possui cinco livros alugados, e se ele atualmente não está num período de infração de 30 dias por falhar em devolver um livro dentro do prazo.

+++ BibliotecaVirtual.Data/Interfaces/IUserRepository.cs
@@ -22,7 +22,7 @@ namespace BibliotecaVirtual.Data.Interfaces
         /// Retorna o UsuarioId do usuário logado.
         /// </summary>
         /// <returns></returns>
-        Task<long?> ObterUsuarioId();
+        Task<int?> ObterUsuarioId();
 
         /// <summary>
         /// Retorna o Nome do usuário logado.
